# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inter Finder API",
    "version": "1.0.0",
    "description": "API Documentation for the Inter Finder web application. This provides a clear contract for frontend and backend teams."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Job": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Reference to Company model"
          },
          "title": {
            "type": "string",
            "example": "Frontend Developer Intern",
            "required": true
          },
          "type": {
            "type": "string",
            "enum": [
              "remote",
              "onsite",
              "hybrid"
            ],
            "required": true
          },
          "location": {
            "type": "string",
            "example": "Remote",
            "required": true
          },
          "salary": {
            "type": "string",
            "example": "$20-$30 per hour",
            "required": true
          },
          "duration": {
            "type": "string",
            "example": "3 months",
            "required": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "deadline": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string",
            "required": true
          },
          "responsibilities": {
            "type": "string",
            "required": true
          },
          "requirements": {
            "type": "string",
            "required": true
          },
          "benefits": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "filled",
              "cancelled"
            ],
            "default": "active"
          },
          "views": {
            "type": "integer",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/applications/company": {
      "get": {
        "summary": "Get all applications for company's listings",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves all applications for listings created by the company",
        "responses": {
          "200": {
            "description": "Successfully retrieved applications"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          }
        }
      }
    },
...
    "/companies/profile": {
      "get": {
        "summary": "Get current company's profile",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved company profile"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          },
          "404": {
            "description": "Company profile not found"
          }
        }
      },
      "put": {
        "summary": "Create or update company profile",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "industry": {
                    "type": "string"
                  },
                  "size": {
                    "type": "string"
                  },
                  "founded": {
                    "type": "number"
                  },
                  "headquarters": {
                    "type": "string"
                  },
                  "logo": {
                    "type": "string"
                  },
                  "coverPhoto": {
                    "type": "string"
                  },
                  "socialMedia": {
                    "type": "object"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "zipCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company profile updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          }
        }
      }
    },
    "/companies": {
      "get": {
        "summary": "Get all companies",
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "select",
            "schema": {
              "type": "string"
            },
            "description": "Fields to select (comma-separated)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Sort by fields (prefix with - for descending)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Company"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "summary": "Get company by ID",
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "404": {
            "description": "Company not found"
          }
        }
      }
    },
...
    "/interviews/company": {
      "get": {
        "summary": "Get all interviews for company's applications",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved interviews"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (company only)"
          }
        }
      }
    },
...
    "/jobs/company/{companyId}": {
      "get": {
        "summary": "Get jobs by company",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs for the company",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}